got rid of load(ed) directory

+++ entrypoint.sh
@@ -3,25 +3,36 @@ set -x
 
 sleep 30
 
-cp /var/www/html/config-docker-compose.ttl /tmp/config-docker-compose.ttl
+if [ -f /data/config.ttl ];then
+    cp /data/config.ttl /tmp/config-docker-compose.ttl   
+else 
+    cp /var/www/html/config-docker-compose.ttl /tmp/config-docker-compose.ttl
+fi
 
-LOADED=`ls /data/loaded/* | wc -l`
-if [ $LOADED -eq 0 ];then
-    curl -u admin:admin -XPOST --data "dbName=skosmos&dbType=tdb" -G http://fuseki:3030/$/datasets
+LOADED=`ls /data/*.loaded | wc -l`
+if [ ${LOADED} -eq 0 ];then
+    curl -u admin:admin -XPOST --data "dbName=skosmos&dbType=tdb" -G ${FUSEKI}/\$/datasets
 fi
 
 PWD=`pwd`
-cd /data/load
+cd /data
 for VOCAB in `ls *.ttl`; do
-    echo "LOAD VOCAB[$VOCAB] ..."
-    GRAPH=`grep sparqlGraph $VOCAB.config | sed 's|    skosmos:sparqlGraph <||g' | sed 's|>$||g'`
-    echo "... in GRAPH[$GRAPH] ..."
-    curl -X PUT -H Content-Type:text/turtle -T $VOCAB -G http://fuseki:3030/skosmos/data --data-urlencode graph=$GRAPH
-    mv $VOCAB $VOCAB.config ../loaded
-    echo "... DONE"
+    VOCAB=`basename ${VOCAB} .ttl`
+    if [ ! -f ${VOCAB}.loaded ]; then
+        echo "LOAD VOCAB[${VOCAB}] ..."
+        GRAPH=`grep sparqlGraph ${VOCAB}.config | sed 's|\s*skosmos:sparqlGraph <\s*||g' | sed 's|\s*>\s*||g'`
+        echo "... in GRAPH[${GRAPH}] ..."
+        if [ ! -f ${VOCAB}.ttl ]; then
+            echo "!ERROR: /data/${VOCAB}.ttl doesn't exist!"
+            exit 1
+        fi
+        curl -X PUT -H Content-Type:text/turtle -T ${VOCAB}.ttl -G ${FUSEKI}/skosmos/data --data-urlencode graph=${GRAPH}
+        touch /data/${VOCAB}.loaded
+        echo "... DONE"
+    fi
 done
-cat /data/loaded/*.ttl.config >>  /tmp/config-docker-compose.ttl
-cd $PWD
+cat /data/*.config >> /tmp/config-docker-compose.ttl
+cd ${PWD}
 
 /usr/bin/envsubst < /tmp/config-docker-compose.ttl > /var/www/html/config.ttl
 cat /var/www/html/config.ttl
